Правки по замечаниям ревьюера 2 часть
