--- src/services/actions/details-ingredient.js
@@ -1,13 +1,0 @@
-import {
-  SET_DETAILS_INGREDIENT,
-  CLEAN_DETAILS_INGREDIENT
-} from '../constants';
-
-export const setDetailsIngredient = (data) => ({
-  type: SET_DETAILS_INGREDIENT,
-  data: data
-});
-
-export const cleanDetailsIngredient = () => ({
-  type: CLEAN_DETAILS_INGREDIENT,
-});
