+++ src/services/reducers/index.js
@@ -3,7 +3,6 @@ import { combineReducers } from 'redux';
 import tabs from './tabs';
 import ingredients from './ingredients';
 import selectedIngredients from './selected-ingredients';
-import detailsIngredient from './details-ingredient';
 import order from './order';
 import auth from './auth';
 
@@ -11,7 +10,6 @@ export default combineReducers({
   tabs,
   ingredients,
   selectedIngredients,
-  detailsIngredient,
   order,
   auth
 });
